We're so close!!
I added to isCPUBoundDone a check to make sure there weren’t any
CPU-bound processes in the CPU (its possible that the final few
CPU-bound processes could be in the CPU but no in the readyQueue or
waitingList, so we’d finish prematurely). I then added an extra tick
right before each algorithm finishes execution in order to print out
the final termination message (the final CPU-bound process would be
terminated, but we wouldn’t get around to doing one more tick inside
the loops because the !(isCPUBoundDone) condition would be false). The
only edit I made in Process was to add in a closing parentheses in
ttlBrstStr().